chore(docs): add migration v8 -> v9 + rename v9 -> v10
