+++ .changeset/chubby-planes-appear.md
@@ -1,0 +1,5 @@
+---
+'@swisspost/design-system-documentation': patch
+---
+
+Added the migration guide for v8 to v9 and updated the migration guide for v9 to v10.
