+++ packages/documentation/src/stories/misc/migration-guide/migration-guide.docs.mdx
@@ -7,6 +7,7 @@ import './migrationv5-6.component';
 import './migrationv6-7.component';
 import './migrationv7-8.component';
 import './migrationv8-9.component';
+import './migrationv9-10.component';
 
 <Meta of={MigrationGuideStories} />
 
@@ -14,18 +15,20 @@ import './migrationv8-9.component';
 
 <migration-global-state>
 
-  ## Setup for migration instructions
+## Setup for migration instructions
 
-  <migration-setup></migration-setup>
+<migration-setup></migration-setup>
 
-  <migration-version-4-5></migration-version-4-5>
+<migration-version-4-5></migration-version-4-5>
 
-  <migration-version-5-6></migration-version-5-6>
+<migration-version-5-6></migration-version-5-6>
 
-  <migration-version-6-7></migration-version-6-7>
+<migration-version-6-7></migration-version-6-7>
 
-  <migration-version-7-8></migration-version-7-8>
+<migration-version-7-8></migration-version-7-8>
 
-  <migration-version-8-9></migration-version-8-9>
+<migration-version-8-9></migration-version-8-9>
+
+<migration-version-9-10></migration-version-9-10>
 
 </migration-global-state>
