+++ packages/documentation/src/stories/misc/migration-guide/migrationv7-8.component.ts
@@ -76,36 +76,6 @@ export class MigrationV78Component extends LitElement {
               </li>
             </ol>
           </li>
-          <li>
-            <h3>Deprecation notice</h3>
-            <p>
-              This is the last major release that supports
-              <a href="https://getbootstrap.com/">Bootstrap framework</a>${this.angular ? html` and
-              <a href="https://ng-bootstrap.github.io/#/home">ng-bootstrap library</a>` : nothing}.
-              Support for these packages will be dropped with v9. Long term support for v8 will be
-              active until the end of 2025 and will receive critical bugfixes.
-            </p>
-            <p>
-              Bootstrap will be removed as a dependency for the Design System. For easy upgrades in
-              the future, the component structure and class names will remain the same, only the
-              styles in the background will be switched to the new token system.
-            </p>
-            <p>
-              Certain utility classes will no longer be avaiable from the Design System due to this
-              change. However, projects that need them can add the bootstrap library as their
-              dependency and conditionally import the needed files. More information on this change
-              will be detailed in the migration guide for v9.
-            </p>
-            ${this.angular ? html`<p>
-               Any ng-bootstrap components will be replaced by web components and available for
-              Angular users with the
-              <a href="https://design-system.post.ch/?path=/docs/833ef689-a573-40f5-a6a6-30a999b94733--docs">
-              Design System Angular Components package</a>
-              . With this change, we will limit our dependency on Angular to the
-              components-angular package which allows us to ship Angular upgrades much faster in the
-              future.
-            </p>` : nothing}
-          </li>
 
           <li>
             <h3>🥮 Rejoice</h3>
